Add signup submit event
