+++ shared/common/src/hooks/index.ts
@@ -1,3 +1,4 @@
 export { default as useModal } from './useModal'
 export { default as useFilterSelect } from './useFilterSelect'
-export { default as  useIntersectionObserver } from './useIntersectionObserver'
+export { default as useIntersectionObserver } from './useIntersectionObserver'
+export { default as useSignUp } from './useSignUp'
